+++ PSVirtualEnv/Functions/ActivatePSVirtualEnv.ps1
@@ -54,34 +54,7 @@ function ValidateVirtualEnvDirectories {
 
 #   function ----------------------------------------------------------------
 # ---------------------------------------------------------------------------
-function ValidateVirtualEnvRequirement {
-
-    <#
-    .DESCRIPTION
-        Return values for the use of validating existing requirement files.
-    
-    .OUTPUTS
-        System.String[]. Requirement files
-    #>
-
-    [CmdletBinding(PositionalBinding)]
-    
-    [OutputType([System.String[]])]
-
-    Param()
-
-    Process{
-
-        $file_list = (Get-ChildItem -Path $PSVirtualEnv.RequireDir -Include "*requirements.txt" -Recurse).FullName
-        return ($file_list | ForEach-Object {
-            $_ -replace ($PSVirtualEnv.RequireDir -replace "\\", "\\")})
-
-    }
-}
-
-#   function ----------------------------------------------------------------
-# ---------------------------------------------------------------------------
-function ValidateVirtualEnvLocalDirectories {
+function ValidateVirtualEnvLocalDirs {
 
     <#
     .DESCRIPTION
@@ -102,5 +75,115 @@ function ValidateVirtualEnvLocalDirectories {
         $file_list = (Get-ChildItem -Path $PSVirtualEnv.LocalDir -Directory)
         return ($file_list | ForEach-Object {
             $_ -replace ($PSVirtualEnv.LocalDir -replace "\\", "\\")})
+    }
+}
+
+
+#   function ----------------------------------------------------------------
+# ---------------------------------------------------------------------------
+function ValidateVirtualEnvSearchDirs {
+    <#
+    .DESCRIPTION
+        Return values for the use of validating existing files with additional search directories.
+    
+    .OUTPUTS
+        System.String[]. Path files
+    #>
+
+    [CmdletBinding(PositionalBinding)]
+    
+    [OutputType([System.String[]])]
+
+    Param()
+
+    Process{
+
+        $file_list = (Get-ChildItem -Path $PSVirtualEnv.SearchDir -Include "*searchdirs.txt" -Recurse).FullName
+        return ($file_list | ForEach-Object {
+            $_ -replace ($PSVirtualEnv.SearchDir -replace "\\", "\\")})
+
+    }
+}
+
+#   function ----------------------------------------------------------------
+# ---------------------------------------------------------------------------
+function ValidateVirtualEnvTemplates {
+    <#
+    .DESCRIPTION
+        Return values for the use of validating existing templates.
+    
+    .OUTPUTS
+        System.String[]. Path files
+    #>
+
+    [CmdletBinding(PositionalBinding)]
+    
+    [OutputType([System.String[]])]
+
+    Param()
+
+    Process{
+
+        return (Get-VirtualEnvFile -Settings -Unformatted | Select-Object -ExpandProperty Name)
+
+    }
+}
+
+#   function ----------------------------------------------------------------
+# ---------------------------------------------------------------------------
+function ValidateVirtualEnvFiles {
+
+    <#
+    .DESCRIPTION
+        Return values for the use of validating existing requirement or script files.
+    
+    .OUTPUTS
+        System.String[]. Requirement or script files
+    #>
+
+    [CmdletBinding(PositionalBinding)]
+    
+    [OutputType([System.String[]])]
+
+    Param(
+
+        [ValidateSet("Requirement", "Script")]
+        [Parameter(Position=1, Mandatory)]
+        [System.String] $Type, 
+        
+        [Switch] $Folder
+    )
+
+    Process{
+
+        switch ($Type) {
+            "Requirement" {
+                $file_path = $PSVirtualEnv.RequireDir
+                $file_include = "*requirements.txt"
+                break; 
+            }
+            "Script" {
+                $file_path = $PSVirtualEnv.ScriptDir
+                $file_include = "*.py", "*.ps1"
+                break;
+            }
+        }
+
+        if (-not $Folder) {1
+            $file_list = (Get-ChildItem -Path $file_path -Include $file_include -Recurse).FullName
+            
+            $result = $file_list | ForEach-Object {
+                $_ -replace ($file_path -replace "\\", "\\")}
+        }
+        else {
+            $folder_list = (Get-ChildItem -Path $file_path -Directory -Recurse).FullName
+
+            $result = $folder_list | ForEach-Object {
+                $_ -replace ($file_path -replace "\\", "\\")
+            } | Where-Object {$_ -notmatch ".git"}
+        }
+        
+        return $result
+
     }
 }