+++ PSVirtualEnv/Functions/Invoke-VirtualEnv.ps1
@@ -41,20 +41,27 @@ function Invoke-VirtualEnv {
         None.
     #>
     
-    [CmdletBinding(PositionalBinding)]
+    [CmdletBinding(PositionalBinding, DefaultParameterSetName="ArgumentList")]
 
     [OutputType([Void])]
 
     Param(
 
         [ValidateSet([ValidateVirtualEnv])]
-        [Parameter(Position=1, ValueFromPipeline, HelpMessage="Name of the virtual environment.")]
+        [Parameter(Position=1, ValueFromPipeline, Mandatory, HelpMessage="Name of the virtual environment.")]
         [System.String] $Name,
 
-        [Parameter(Position=2, HelpMessage="Specifies parameters or parameter values to use when this cmdlet starts the process. If parameters or parameter values contain a space, they need surrounded with escaped double quotes.")]
+        [Parameter(ParameterSetName="ArgumentList", Position=2, Mandatory,  HelpMessage="Specifies parameters or parameter values to use when this cmdlet starts the process. If parameters or parameter values contain a space, they need surrounded with escaped double quotes.")]
         [System.String[]] $ArgumentList,
 
-        [Parameter(HelpMessage="Specifies the commands to run. Enclose the commands in braces ( { } ) to create a script block.")]
+        [ValidateSet([ValidateVenvScripts])]
+        [Parameter(ParameterSetName="Script", Position=2, ValueFromPipeline, HelpMessage="Relative path to a script file in predefined scripts folder.")]
+        [System.String] $Script,
+
+        [Parameter(ParameterSetName="Script", Position=3, ValueFromPipeline, HelpMessage="Specifies parameters or parameter values to use when this cmdlet starts the process. If parameters or parameter values contain a space, they need surrounded with escaped double quotes.")]
+        [System.String] $ScriptArgumentList,
+
+        [Parameter(ParameterSetName="ScriptBlock", Position=2, HelpMessage="Specifies the commands to run. Enclose the commands in braces ( { } ) to create a script block.")]
         [ScriptBlock] $ScriptBlock,
 
         [Parameter(HelpMessage="If switch 'silent' is true no output will written to host.")]
@@ -63,21 +70,37 @@ function Invoke-VirtualEnv {
 
     Process{
 
-        Set-VirtualEnv -Name $Name
-        
-        $invokation_cmd = $ScriptBlock
-        if ($ArgumentList) {
-            $invokation_cmd = [ScriptBlock]::Create($(Get-ArgumentList -ArgumentList $ArgumentList))
+        Start-VirtualEnv -PrivateName $Name -Silent
+
+        switch ($PSCmdlet.ParameterSetName) {
+            "ArgumentList" {
+                $script_array =  $ArgumentList
+                break;
+            }
+            "Script" {
+                # get existing python script file 
+                $script_file = Join-Path -Path $PSVirtualEnv.ScriptDir -ChildPath $Script
+                
+                if ($Script -match "[^\.].py$"){
+                    $script_array = @("python", $script_file) + $ScriptArguments
+                } else {
+                    $script_array= @("." , $script_file) + $ScriptArguments
+                }
+
+                break;
+            }
         }
 
+        if ($PSCmdlet.ParameterSetName -ne "ScriptBlock"){
+            $ScriptBlock = [ScriptBlock]::Create($(Get-ArgumentList -ArgumentList $script_array))
+        }
 
         if ($Silent) {
-            $invokation_cmd.Invoke() 2>&1> $Null
+            $ScriptBlock.Invoke() 2>&1> $Null
         } else {
-            $invokation_cmd.Invoke()
+            $ScriptBlock.Invoke()
         }
 
-        Restore-VirtualEnv
-
+        Stop-VirtualEnv -Silent
     }
 }